NH0098

Conjunto de datos recopilados en la estacion NH0098 del INTA

Usage

NH0098

Format

id

Clave unica de identificacion de la estacion

fecha

Fecha de la observacion

temperatura_abrigo_150cm

Temperatura medida a 150 cm del suelo, en grados Celsius.

temperatura_abrigo_150cm_maxima

Temperatura maxima medida a 150 cm del suelo, en grados Celsius.

temperatura_abrigo_150cm_minima

Temperatura minima medida a 150 cm del suelo, en grados Celsius.

temperatura_intemperie_5cm_minima

Temperatura minima medida al aire libre a 5 cm del suelo, en grados Celsius.

temperatura_intemperie_50cm_minima

Temperatura minima medida al aire libre a 50 cm del suelo, en grados Celsius.

temperatura_suelo_5cm_media

Temperatura media del suelo a 5cm de profundidad, en grados Celsius.

temperatura_suelo_10cm_media

Temperatura media del suelo a 10 cm de profundidad, en grados Celsius.

temperatura_inte_5cm

Temperatura medida al aire libre a 5 cm del suelo, en grados Celsius.

temperatura_intemperie_150cm_minima

Temperatura minima medida al aire libre a 150 cm del suelo, en grados Celsius.

humedad_suelo

Humedad medida en el suelo.

precipitacion_pluviometrica

Cantidad de precipitacion acumulada, en milimetros.

granizo

1: Si, 0:No.

nieve

1: Si, 0:No.

heliofania_efectiva

Horas de sol efectivo al dia.

heliofania_relativa

Porcentaje de sol efectivo respecto a las horas de luz disponibles.

tension_vapor_media

Tension de vapor promedio, en hectopascales.

humedad_media

Humedad relativa promedio.

humedad_media_8_14_20

Humedad relativa medida a las 8, 14 y 20 horas.

rocio_medio

Punto de rocio medio

duracion_follaje_mojado

Horas de follaje mojado por dia.

velocidad_viento_200cm_media

Velocidad media del viento a 200 cm del suelo, en km por hora.

direccion_viento_200cm

Direccion del viento a 200 cm del suelo, en grados.

velocidad_viento_1000cm_media

Velocidad media del viento a 1000 cm del suelo, en km por hora.

direccion_viento_1000cm

Direccion del viento a 1000 cm del suelo, en grados.

velocidad_viento_maxima

Velocidad maxima del viento regristrada, en km por hora.

presion_media

Presion atmosferica media, en hectopascales.

radiacion_global

Cantidad de radiacion solar global.

radiacion_neta

Cantidad de radiacion solar neta.

evaporacion_tanque

Evaporacion de agua diaria medida en un tanque, en milimetros,

evapotranspiracion_potencial

Evapotranspiracion potencial diaria, en milimetros .

profunidad_napa

Profunidad de la napa freatica, en metros.

horas_frio

Cantidad de horas de frio acumuladas.

unidad_frio

Unidades de frio acumuladas.

Source

Datos que diponibiliza el Institulo Nacional de Tecnología Agropecuaria INTA a traves del Sistema de Información y Gestión Agropecuaria SIGA.