Lectura de archivos

La función lectura_estaciones() toma como parametro el id de la estación a estudiar y la ruta del archivo, si el archivo ya estaba descargado, lo disponibiliza para su posterior analisis, y de lo contrario, lo descarga y luego lo diaponibiliza

Usage

lectura_estaciones(id_estacion, ruta)

Arguments

id_estacion

conjunto de números y letras que identifica la estación

ruta

la ruta donde se encuentra el archivo

Value

variable que contiene el archivo disponible para su análisis

Examples

lectura_estaciones("NH0437", "~/prueba_paquete")
#> Warning: One or more parsing issues, call `problems()` on your data frame for details,
#> e.g.:
#>   dat <- vroom(...)
#>   problems(dat)
#> Rows: 18929 Columns: 35
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ","
#> chr   (2): id, direccion_viento_1000cm
#> dbl  (22): temperatura_abrigo_150cm, temperatura_abrigo_150cm_maxima, temper...
#> lgl  (10): temperatura_suelo_10cm_media, temperatura_inte_5cm, temperatura_i...
#> date  (1): fecha
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
#> El archivo no existía, por lo cual se descargó y luego se leyó.
#> # A tibble: 18,929 × 35
#>    id     fecha      temperatura_abrigo_150cm temperatura_abrigo_150cm_maxima
#>    <chr>  <date>                        <dbl>                           <dbl>
#>  1 NH0437 1970-08-01                    23.3                             31.2
#>  2 NH0437 1970-08-02                    26.6                             32.1
#>  3 NH0437 1970-08-03                    17.4                             22.9
#>  4 NH0437 1970-08-04                    11.0                             11.9
#>  5 NH0437 1970-08-05                    11.1                             12.6
#>  6 NH0437 1970-08-06                    13.6                             15.1
#>  7 NH0437 1970-08-07                    10.8                             14.9
#>  8 NH0437 1970-08-08                     7.75                            13.7
#>  9 NH0437 1970-08-09                     8.2                             15.1
#> 10 NH0437 1970-08-10                    10.2                             18.4
#> # ℹ 18,919 more rows
#> # ℹ 31 more variables: temperatura_abrigo_150cm_minima <dbl>,
#> #   temperatura_intemperie_5cm_minima <dbl>,
#> #   temperatura_intemperie_50cm_minima <dbl>,
#> #   temperatura_suelo_5cm_media <dbl>, temperatura_suelo_10cm_media <lgl>,
#> #   temperatura_inte_5cm <lgl>, temperatura_intemperie_150cm_minima <lgl>,
#> #   humedad_suelo <lgl>, precipitacion_pluviometrica <dbl>, granizo <dbl>, …
lectura_estaciones("NH0098", "~/seguimos_probando")
#> Warning: One or more parsing issues, call `problems()` on your data frame for details,
#> e.g.:
#>   dat <- vroom(...)
#>   problems(dat)
#> Rows: 23733 Columns: 35
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ","
#> chr   (2): id, direccion_viento_1000cm
#> dbl  (21): temperatura_abrigo_150cm, temperatura_abrigo_150cm_maxima, temper...
#> lgl  (11): temperatura_intemperie_50cm_minima, temperatura_inte_5cm, tempera...
#> date  (1): fecha
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
#> El archivo no existía, por lo cual se descargó y luego se leyó.
#> # A tibble: 23,733 × 35
#>    id     fecha      temperatura_abrigo_150cm temperatura_abrigo_150cm_maxima
#>    <chr>  <date>                        <dbl>                           <dbl>
#>  1 NH0098 1956-04-01                     13.6                            21.6
#>  2 NH0098 1956-04-02                     16.8                            23.6
#>  3 NH0098 1956-04-03                     19                              25.9
#>  4 NH0098 1956-04-04                     17.6                            23.6
#>  5 NH0098 1956-04-05                     20                              27.2
#>  6 NH0098 1956-04-06                     19.2                            24.7
#>  7 NH0098 1956-04-07                     15                              19.1
#>  8 NH0098 1956-04-08                     18.8                            26.8
#>  9 NH0098 1956-04-09                     18.2                            26.6
#> 10 NH0098 1956-04-10                     11.4                            15.4
#> # ℹ 23,723 more rows
#> # ℹ 31 more variables: temperatura_abrigo_150cm_minima <dbl>,
#> #   temperatura_intemperie_5cm_minima <dbl>,
#> #   temperatura_intemperie_50cm_minima <lgl>,
#> #   temperatura_suelo_5cm_media <dbl>, temperatura_suelo_10cm_media <dbl>,
#> #   temperatura_inte_5cm <lgl>, temperatura_intemperie_150cm_minima <lgl>,
#> #   humedad_suelo <lgl>, precipitacion_pluviometrica <dbl>, granizo <dbl>, …