Lectura de archivos

La función lectura_estaciones() toma como parametro el id de la estación a estudiar y la ruta del archivo, si el archivo ya estaba descargado, lo disponibiliza para su posterior analisis, y de lo contrario, lo descarga y luego lo diaponibiliza

Usage

lectura_estaciones(id_estacion, ruta)

Arguments

id_estacion

conjunto de números y letras que identifica la estación

ruta

la ruta donde se encuentra el archivo

Value

variable que contiene el archivo disponible para su análisis

Examples

lectura_estaciones("NH0437", "./documentos/programacion/datos")
#> Warning: URL https://raw.githubusercontent.com/rse-r/intro-programacion/main/datos/NH0437.csv: cannot open destfile './documentos/programacion/datos', reason 'No such file or directory'
#> Warning: download had nonzero exit status
#> Error: './documentos/programacion/datos' does not exist in current working directory ('/home/runner/work/datosclima/datosclima/docs/reference').
lectura_estaciones("NH0098", "./escritorio/programacion/ejs_parcial")
#> Warning: URL https://raw.githubusercontent.com/rse-r/intro-programacion/main/datos/NH0098.csv: cannot open destfile './escritorio/programacion/ejs_parcial', reason 'No such file or directory'
#> Warning: download had nonzero exit status
#> Error: './escritorio/programacion/ejs_parcial' does not exist in current working directory ('/home/runner/work/datosclima/datosclima/docs/reference').